# internal/application/dto/user/change_password_req.go
package userdto

type ChangePasswordReq struct {
	OldPassword        string `json:"old_password"`
	NewPassword        string `json:"new_password"`
	NewPasswordConfirm string `json:"new_password_confirm"`
}